Omit signatures from transaction hash
This allows us to know a transaction hash (ID) before
the signature is attached. This is important to allow
a full node (or a wallet node) to generate a baseline
transaction before attaching the signature.

Trying this change in preparation for supporting a
2-party payment model where one (wallet) node would request
an unsigned transaction from another, attach the signature
by signing with its own public key, and then send it back
in order to get it into the transaction pool.