add unmined block endpoint
miner-only nodes can use this endpoint to request
a structurally-valid but un-mined block that they could
then mine and return to the full node.

miners can submit a DER key to serve as the address
for the coinbase transaction in the requested block, or if they
don't send one the node will use its own.