Exclude Pending Txn Sources when Making Payment
When we produce a payment from one address to another
we should avoid sourcing any inputs that are already
consumed by an input in the pending transaction pool.

Without accounting for this, generating 2 payments within
the timeframe of the same block would result in trying to
re-use the same inputs (since they appear unspent from
the perspective of the confirmed blocks) within that block.

This way an address can continue to make multiple payments
within the same block as long as they have enough funds